Revert LogClient dir rename
