+++ lib/hooks/LogClient/logClient.test.ts
